fix(api): extract pairedSibling resolver to keep definition.ts under 700 lines
